# Conflicts:
#	vpr/src/draw/draw_basic.cpp

+++ vpr/src/route/rr_graph2.cpp
@@ -1453,7 +1453,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                             VPR_ERROR(VPR_ERROR_ROUTE, "RR node type does not match between rr_nodes and rr_node_indices (%s/%s): %s",
                                       rr_node_typename[rr_graph.node_type(inode)],
                                       rr_node_typename[rr_type],
-                                      describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                                      describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
                         }
 
                         if (rr_graph.node_type(inode) == CHANX) {
@@ -1463,7 +1463,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                                 VPR_ERROR(VPR_ERROR_ROUTE, "RR node y position does not agree between rr_nodes (%d) and rr_node_indices (%d): %s",
                                           rr_graph.node_ylow(inode),
                                           y,
-                                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
                             }
 
                             if (!rr_graph.x_in_node_range(x, inode)) {
@@ -1471,7 +1471,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                                           rr_graph.node_xlow(inode),
                                           rr_graph.node_xlow(inode),
                                           x,
-                                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
                             }
                         } else if (rr_graph.node_type(inode) == CHANY) {
                             VTR_ASSERT_MSG(rr_graph.node_xlow(inode) == rr_graph.node_xhigh(inode), "CHANY should be veritcal");
@@ -1480,7 +1480,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                                 VPR_ERROR(VPR_ERROR_ROUTE, "RR node x position does not agree between rr_nodes (%d) and rr_node_indices (%d): %s",
                                           rr_graph.node_xlow(inode),
                                           x,
-                                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
                             }
 
                             if (!rr_graph.y_in_node_range(y, inode)) {
@@ -1488,7 +1488,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                                           rr_graph.node_ylow(inode),
                                           rr_graph.node_ylow(inode),
                                           y,
-                                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
                             }
                         } else if (rr_graph.node_type(inode) == SOURCE || rr_graph.node_type(inode) == SINK) {
                             //Sources have co-ordintes covering the entire block they are in
@@ -1497,7 +1497,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                                           rr_graph.node_xlow(inode),
                                           rr_graph.node_xlow(inode),
                                           x,
-                                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
                             }
 
                             if (!rr_graph.y_in_node_range(y, inode)) {
@@ -1505,7 +1505,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                                           rr_graph.node_ylow(inode),
                                           rr_graph.node_ylow(inode),
                                           y,
-                                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
                             }
 
                         } else {
@@ -1568,7 +1568,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                           rr_area,
                           rr_node.length(),
                           count,
-                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
             }
             /* As we allow a pin to be indexable on multiple sides,
              * This check code should not be applied to input and output pins
@@ -1578,7 +1578,7 @@ bool verify_rr_node_indices(const DeviceGrid& grid,
                 VPR_ERROR(VPR_ERROR_ROUTE, "Mismatch between RR node length (%d) and count within rr_node_indices (%d, should be length + 1): %s",
                           rr_node.length(),
                           count,
-                          describe_rr_node(rr_graph, grid, rr_indexed_data, size_t(inode), is_flat).c_str());
+                          describe_rr_node(rr_graph, grid, rr_indexed_data, inode, is_flat).c_str());
             }
         }
     }
